Don't leak event listeners
`f.bind(x)` won't return the same bound context between calls, so the
`.off` can't find the same listener and disconnect it, so we leak
listeners and potentially other closed contexts.

PusherStore doesn't rebind event listeners anyway so we should be okay
without these binds, as long as we access the prototyped listeners.